Use flag.Int for ntpQueryTimeoutMilliseconds.

+++ main.go
@@ -25,7 +25,7 @@ var (
 	filterDNSToIPV4Only         = flag.Bool("filterDNSToIPV4Only", true, "filter DNS results to IPv4 only")
 	maxParallelDNSRequests      = flag.Int("maxParallelDNSRequests", 2, "max parallel DNS requests")
 	maxParallelNTPRequests      = flag.Int("maxParallelNTPRequests", 8, "max parallel NTP requests")
-	ntpQueryTimeoutMilliseconds = flag.Int64("ntpQueryTimeoutMilliseconds", 1_000, "NTP query timeout milliseconds")
+	ntpQueryTimeoutMilliseconds = flag.Int("ntpQueryTimeoutMilliseconds", 1_000, "NTP query timeout milliseconds")
 	slogLevel                   slog.Level
 )
